requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
dotenv==0.9.9
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
